bpf: Merge bpf_lb and bpf_netdev
Move load balancing functionality from bpf_lb into bpf_netdev. This will
allow for both the host and external network endpoints to reach Cilium
service IPs.

Also ensures that we *always* return TC_ACT_OK on a parsing failure to
never risk making the host become unavailable.
