+++ bpf/node_config.h
@@ -51,3 +51,5 @@
 #define TUNNEL_ENDPOINT_MAP_SIZE 65536
 #define ENDPOINTS_MAP_SIZE 65536
 #define ENABLE_TRACE
+#define LB_IP4
+#define LB_IP6
