+++ daemon/bpf.sha
@@ -1,2 +1,2 @@
-GO_BINDATA_SHA1SUM=968d3b81a6586cb7ae1c3a1f638db7f8acc5b81e
+GO_BINDATA_SHA1SUM=5e46a10e0a7b5f4c3cdadd9b898433e71f11ba7f
 GO_VERSION_USED=go1.8.3
